fixed size of node hash
The node hash was 1 item too small and therefor it was overwriting the last item with other data.
I think it will crash less often now.
But there are still some tables that need to be increased and more debugging need to be done aswell to see if everything really works as it should.