Added some logic to the hybrid stutter fix
It seems on some hardware it might make cpu slightly slower and produce slightly less than 60 fps, specially if in fullscreen for some reason.
Now with the logic it should always be atleast 60 fps, however on worst case scenario it will not sleep anything so basically exact should be used instead.
To test this go into debug mode and see if Use exact is printed.