fixed #102 and #80
The problem was that timers were depending on how long your computer had been on.
Th3 fixed this by "reseting" the timers at beginning of application.

Now timers are reset every time you change level and the issues should not be able to happen unless you are in the same level for a VERY long time